Updated the team page, error page, and the SSI files
I updated the pages I worked on (Team and Error) so they look as well as they did when they were still HTML files.  In addition, I added new feats to the team page, got both of my pages to use the header and footer SSI files, and updated said SSI files with the code we all share in the headers and footers of our pages.