Updated member.php
One of the issues of the old version of member of php is that it wouldn't run because I messed up an IF statement.  This version of member.php doesn't have that issue and goes through both the cookies from login and register to determine if the member is in or not.  InvalidLogin == "true" didn't want to cooperate when I tested it though.