Updated Flavor Text
I updated the flavor text of each of the files I worked on.  While error and TeamError weren't given much of an update, team has been updated with an increased description for Marielle and member now displays the username in the title.  Now all that's left to do is figure out how backgrounds work.  My gut's telling me it's a size issue.